chore: Refactor File format
This commit removes reserved allocations that adds on unwanted space.

Additionally:
1. Some files were refactored for better readability.
2. File format byte sizes were modified.

Next Work:
1. Apply validations on the Database and Table layer
2. Move all CLI outputs to the main file instead of the database module
3. Apply clean wizard UI's

+++ src/database/model.rs
@@ -24,7 +24,7 @@ pub enum DataType {
 }
 
 impl DataType {
-    pub fn id(&self) -> u32 {
+    pub fn id(&self) -> u8 {
         match self {
             DataType::Integer => 0,
             DataType::Float => 1,
